add placer (for placing stuff)
